Make test actually test the issue that was fixed, add basic caching for reexports
