+++ test/chunking-form/samples/preserve-modules-nested-export/_config.js
@@ -1,7 +1,7 @@
 module.exports = {
 	description: 'confirm exports are preserved when exporting a module',
 	options: {
-		input: ['main1.js', 'main2.js'],
+		input: 'main.js',
 		preserveModules: true
 	}
 };
