--- test/chunking-form/samples/preserve-modules-nested-export/_expected/amd/dep.js
@@ -1,7 +1,0 @@
-define(['exports'], function (exports) { 'use strict';
-
-	const foo = 1;
-
-	exports.foo = foo;
-
-});
