+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/amd/inner/more_inner/something.js
@@ -1,0 +1,7 @@
+define(['exports'], function (exports) { 'use strict';
+
+	const Something = 'Hello World';
+
+	exports.Something = Something;
+
+});
