+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/amd/main.js
@@ -1,0 +1,9 @@
+define(['exports', './inner/more_inner/something'], function (exports, __chunk_1) { 'use strict';
+
+
+
+	exports.Something = __chunk_1.Something;
+
+	Object.defineProperty(exports, '__esModule', { value: true });
+
+});
