--- test/chunking-form/samples/preserve-modules-nested-export/_expected/amd/main2.js
@@ -1,9 +1,0 @@
-define(['exports', './dep'], function (exports, __chunk_1) { 'use strict';
-
-
-
-	exports.foo = __chunk_1.foo;
-
-	Object.defineProperty(exports, '__esModule', { value: true });
-
-});
