--- test/chunking-form/samples/preserve-modules-nested-export/_expected/cjs/dep.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const foo = 1;
-
-exports.foo = foo;
