+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/cjs/inner/more_inner/something.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const Something = 'Hello World';
+
+exports.Something = Something;
