+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/cjs/main.js
@@ -1,0 +1,9 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var __chunk_1 = require('./inner/more_inner/something.js');
+
+
+
+exports.Something = __chunk_1.Something;
