--- test/chunking-form/samples/preserve-modules-nested-export/_expected/cjs/main2.js
@@ -1,9 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', { value: true });
-
-var __chunk_1 = require('./dep.js');
-
-
-
-exports.foo = __chunk_1.foo;
