--- test/chunking-form/samples/preserve-modules-nested-export/_expected/es/dep.js
@@ -1,3 +1,0 @@
-const foo = 1;
-
-export { foo };
