+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/es/inner/more_inner/something.js
@@ -1,0 +1,3 @@
+const Something = 'Hello World';
+
+export { Something };
