+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/es/main.js
@@ -1,0 +1,1 @@
+export { Something } from './inner/more_inner/something.js';
