--- test/chunking-form/samples/preserve-modules-nested-export/_expected/es/main2.js
@@ -1,1 +1,0 @@
-export { foo } from './dep.js';
