+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/system/inner/more_inner/something.js
@@ -3,7 +3,7 @@ System.register([], function (exports, module) {
 	return {
 		execute: function () {
 
-			const foo = exports('foo', 1);
+			const Something = exports('Something', 'Hello World');
 
 		}
 	};
