+++ test/chunking-form/samples/preserve-modules-nested-export/_expected/system/main.js
@@ -1,0 +1,13 @@
+System.register(['./inner/more_inner/something.js'], function (exports, module) {
+	'use strict';
+	return {
+		setters: [function (module) {
+			exports('Something', module.Something);
+		}],
+		execute: function () {
+
+
+
+		}
+	};
+});
