--- test/chunking-form/samples/preserve-modules-nested-export/_expected/system/main1.js
@@ -1,13 +1,0 @@
-System.register(['./dep.js'], function (exports, module) {
-	'use strict';
-	return {
-		setters: [function (module) {
-			exports('foo', module.foo);
-		}],
-		execute: function () {
-
-
-
-		}
-	};
-});
