+++ test/chunking-form/samples/preserve-modules-nested-export/inner/index.js
@@ -1,0 +1,1 @@
+export * from './more_inner/index';
