+++ test/chunking-form/samples/preserve-modules-nested-export/inner/more_inner/index.js
@@ -1,0 +1,1 @@
+export * from './something';
