+++ test/chunking-form/samples/preserve-modules-nested-export/inner/more_inner/something.js
@@ -1,0 +1,1 @@
+export const Something = 'Hello World';
