+++ test/chunking-form/samples/preserve-modules-nested-export/main.js
@@ -1,0 +1,1 @@
+export * from './inner/index';
