--- test/chunking-form/samples/preserve-modules-nested-export/main1.js
@@ -1,1 +1,0 @@
-export { foo } from './dep';
