--- test/chunking-form/samples/preserve-modules-nested-export/main2.js
@@ -1,1 +1,0 @@
-export { foo } from './main1.js';
